Create count_elements_in_defaultdict.py
The code in this script can be used to count elements in a dictionary - sometimes we need to apply different operations on the total counts or fractions of the elements that pass a certain criteria. This script includes several methods that can be useful to access, count and output results on certain members of the dictionary.